+++ airflow/docker-compose.yaml
@@ -71,6 +71,7 @@ x-airflow-common: &airflow-common
     volumes:
         - ${AIRFLOW_PROJ_DIR:-.}/dags:/opt/airflow/dags
         - ${AIRFLOW_PROJ_DIR:-.}/logs:/opt/airflow/logs
+        # - ${AIRFLOW_PROJ_DIR:-.}/tasks:/opt/airflow/tasks
         - ${AIRFLOW_PROJ_DIR:-.}/plugins:/opt/airflow/plugins
         - ${AIRFLOW_PROJ_DIR:-.}/credentials:/opt/airflow/credentials
         - ${AIRFLOW_PROJ_DIR:-.}/../../data_engineer_test_v2:/opt/airflow/files
